profiling-authentication/requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
pip==20.2.3
psycopg2==2.8.6
pytz==2020.1
setuptools==50.3.0
sqlparse==0.3.1
